+++ core/src/bms/player/beatoraja/song/SongData.java
@@ -157,7 +157,9 @@ public class SongData implements Validatable, IpfsInformation {
 		banner = model.getBanner();
 
 		setStagefile(model.getStagefile());
-		setBackbmp(model.getBackbmp());
+		if (backbmp == null || backbmp.isEmpty()) {
+			setBackbmp(model.getBackbmp());
+		}
         if(preview == null || preview.length() == 0) {
             setPreview(model.getPreview());
         }